05_logging_monitoring/func01/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
requests
opencensus-extension-azure-functions
opencensus-ext-requests